Add files via upload
Space Invaders is a two-dimensional shooter game in which the player controls a laser cannon by moving it horizontally across the bottom of the screen and firing at descending aliens. The aim is to defeat aliens.
The game is controlled with arrow left, arrow right and b of the keyboard to move the player left, right and fire bullets respectively.
At the beginning of the game, the health of a player is 3. Each time the spacecraft of player is hit by the bullet of Aliens, the health of a player decreases by 1.
The score of the player increases by 5 each time it kills the alien.